requirements.txt
PyJWT
requests
python-dotenv
cryptography
cffi
pyyaml